update with rebased changes applied to correct manager
