+++ controllers/pipelines/provider_deployment_manager.go
@@ -2,6 +2,7 @@ package pipelines
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	config "github.com/sky-uk/kfp-operator/apis/config/v1alpha6"
 	. "github.com/sky-uk/kfp-operator/apis/pipelines"
@@ -10,6 +11,7 @@ import (
 	"github.com/sky-uk/kfp-operator/controllers/pipelines/predicates"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -18,6 +20,7 @@ import (
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log"
+	"strings"
 )
 
 type DeploymentResourceManager interface {
@@ -140,16 +143,35 @@ func populateServiceContainer(serviceContainerName string, podTemplate corev1.Po
 		return nil, fmt.Errorf("unable to populate service container: container with name %s not found on deployment", serviceContainerName)
 	}
 
+	paramsAsEnvVars := make([]corev1.EnvVar, len(provider.Spec.Parameters))
+	for name, value := range provider.Spec.Parameters {
+		paramsAsEnvVars = append(paramsAsEnvVars, corev1.EnvVar{Name: fmt.Sprintf("PARAMETERS_%s", strings.ToUpper(name)), Value: jsonToString(value)})
+	}
+
 	podTemplate.Spec.Containers = Map(podTemplate.Spec.Containers, func(c corev1.Container) corev1.Container {
 		if c.Name == serviceContainerName {
 			c.Image = provider.Spec.ServiceImage
-			c.Env = append(c.Env, corev1.EnvVar{
-				Name:  "PROVIDERNAME",
-				Value: provider.Name,
-			})
+			c.Env = append(c.Env,
+				corev1.EnvVar{
+					Name:  "PROVIDERNAME",
+					Value: provider.Name,
+				},
+			)
+			c.Env = append(c.Env, paramsAsEnvVars...)
 		}
 		return c
 	})
 
 	return &podTemplate, nil
+}
+
+func jsonToString(jsonValue *apiextensionsv1.JSON) string {
+	var s string
+	if jsonValue != nil {
+		// Attempts to unmarshal input into a string to remove extra quotes and escape chars if the input is, in fact, a string
+		if err := json.Unmarshal(jsonValue.Raw, &s); err != nil {
+			s = string(jsonValue.Raw)
+		}
+	}
+	return s
 }
